requirements.txt
matplotlib~=3.7.1
numpy~=1.24.2
scipy~=1.10.1
tensorflow~=2.11.0
yaml~=0.2.5
pyyaml~=6.0
imageio~=2.26.0
pillow~=9.4.0
tf_slim~=1.1.0
colorama~=0.4.6